Trekk ut mock-handlere i egne filer
