--- src/mocks/arbeidsliste.ts
@@ -1,16 +1,0 @@
-export default [
-    {
-        sistEndretAv: {
-            veilederId: 'TestVeilederID1'
-        },
-        endringstidspunkt: '2022-01-10T10:09:23.497214+01:00',
-        overskrift: 'Arbeidsliste 1',
-        kommentar: 'Test Kommentar 1 ',
-        frist: '2022-01-15T00:00:00+01:00',
-        isOppfolgendeVeileder: null,
-        arbeidslisteAktiv: null,
-        kategori: 'LILLA',
-        harVeilederTilgang: null,
-        aktoerid: '1234'
-    }
-];
