+++ src/mocks/data/mine-filter.ts
@@ -1,6 +1,6 @@
-import {initialState} from '../ducks/filtrering';
+import {initialState} from '../../ducks/filtrering';
 import {faker} from '@faker-js/faker/locale/nb_NO';
-import {LagretFilter} from '../ducks/lagret-filter';
+import {LagretFilter} from '../../ducks/lagret-filter';
 
 export const mineFilter = () => {
     return [
