+++ src/mocks/data/portefolje.ts
@@ -1,8 +1,8 @@
 import {innloggetVeileder, veiledere} from './veiledere';
-import {aktiviteter, hendelserLabels} from '../filtrering/filter-konstanter';
-import {MOCK_CONFIG, rnd} from './utils';
+import {aktiviteter, hendelserLabels} from '../../filtrering/filter-konstanter';
+import {MOCK_CONFIG, rnd} from '../utils';
 import {faker} from '@faker-js/faker/locale/nb_NO';
-import {BarnUnder18Aar, KategoriModell} from '../model-interfaces';
+import {BarnUnder18Aar, KategoriModell} from '../../model-interfaces';
 import moment from 'moment';
 
 faker.seed(MOCK_CONFIG.seed);
