+++ src/mocks/data/portefoljestorrelser.ts
@@ -1,5 +1,5 @@
 import {innloggetVeileder, lagTilfeldigVeilederId} from './veiledere';
-import {rnd} from './utils';
+import {rnd} from '../utils';
 
 function lagPortefoljeStorrelse() {
     return {
